+++ src/graysvr/graysvr.h
@@ -1436,7 +1436,7 @@ public:
 
 
 	void addShowDamage( int damage, DWORD uid_damage );
-	void addSpeedMode( int speedMode = 0 );
+	void addSpeedMode( BYTE speedMode = 0 );
 	void addVisualRange( BYTE visualRange = UO_MAP_VIEW_SIZE );
 	void addIdleWarning( BYTE message );
 	void addKRToolbar( bool bEnable );
